use Qt integers instead of c style integers

+++ animationaudio.h
@@ -4,22 +4,23 @@
 #include "animation.h"
 #include <QFile>
 #include <QTextStream>
+#include <QtGlobal>
 
 #define TWOPI (double)(6.283185307179586476925286766559)
 
 class AnimationAudio
 {
 public:
-	AnimationAudio(Animation* animation, int frameRate, int len);
+	AnimationAudio(Animation* animation, qint32 frameRate, qint32 len);
 	bool generateFile(QString& fileName);
 
 private:
 	Animation*		mAnimation;
 	QFile*			mAudioFile;
 	QTextStream*	mStream;
-	int				mFrameRate;
-	int				mLength;
-	const uint32_t  mSampleRate = 48000;
+    qint16			mFrameRate;
+    qint16			mLength;
+    const quint32   mSampleRate = 48000;
 
 	bool writeHeader(QDataStream* stream);
 	bool writeData(QDataStream* stream);
